Removed deprecated JSONResourceType and NetworkJSONResourceType
