+++ Tests/TABResourceLoaderTests/Mocks/MockDefaultNetworkDataResource.swift
@@ -1,0 +1,21 @@
+//
+//  MockDefaultNetworkDataResource.swift
+//  TABResourceLoader
+//
+//  Created by Luciano Marisi on 02/11/2016.
+//  Copyright © 2016 Luciano Marisi. All rights reserved.
+//
+
+import Foundation
+@testable import TABResourceLoader
+
+protocol NetworkDataResourceType: NetworkResourceType, DataResourceType {}
+
+struct MockDefaultNetworkDataResource: NetworkDataResourceType {
+  typealias Model = String
+  let url: URL
+  
+  func result(from data: Data) -> Result<String> {
+    return .success("")
+  }
+}
